chore: update key derivation method to OpenSSL version of HKDF-SHA256
In the demo example of M-SIDH handshake, a simple SHA256 was used as a key derivation. Given that is is only a demo it does not have that much of an impact, but to make it a bit more mature, key derivation was replaced with HKDF which was designed for this purpose.