refactor: change type of tors_basis order field `n` from pprod_t to mpz_t
Torsion basis was mostly used in the code with n->value accessor. Variable n was required as pprod_t only for isogeny calculation - current form was used to reduce the number of arguments passed to the function (isogeny degree was part of the torsion basis). To make the code easier and more flexible to future changes the type was change.